Split workflow-step into multiple modules
