+++ packages/twenty-server/src/engine/core-modules/workflow/resolvers/workflow-version-step.resolver.ts
@@ -21,7 +21,6 @@ import { PermissionsGraphqlApiExceptionFilter } from 'src/engine/metadata-module
 import { WorkflowVersionStepWorkspaceService } from 'src/modules/workflow/workflow-builder/workflow-version-step/workflow-version-step.workspace-service';
 import { WorkflowActionType } from 'src/modules/workflow/workflow-executor/workflow-actions/types/workflow-action.type';
 import { WorkflowRunWorkspaceService } from 'src/modules/workflow/workflow-runner/workflow-run/workflow-run.workspace-service';
-import { CreateWorkflowVersionEdgeInput } from 'src/engine/core-modules/workflow/dtos/create-workflow-version-edge-input.dto';
 import { WorkflowVersionStepChangesDTO } from 'src/engine/core-modules/workflow/dtos/workflow-version-step-changes.dto';
 
 @Resolver()
@@ -35,7 +34,7 @@ import { WorkflowVersionStepChangesDTO } from 'src/engine/core-modules/workflow/
   PermissionsGraphqlApiExceptionFilter,
   PreventNestToAutoLogGraphqlErrorsFilter,
 )
-export class WorkflowStepResolver {
+export class WorkflowVersionStepResolver {
   constructor(
     private readonly workflowVersionStepWorkspaceService: WorkflowVersionStepWorkspaceService,
     private readonly workflowRunWorkspaceService: WorkflowRunWorkspaceService,
@@ -122,33 +121,5 @@ export class WorkflowStepResolver {
     });
 
     return step;
-  }
-
-  @Mutation(() => WorkflowVersionStepChangesDTO)
-  async createWorkflowVersionEdge(
-    @AuthWorkspace() { id: workspaceId }: Workspace,
-    @Args('input')
-    { source, target, workflowVersionId }: CreateWorkflowVersionEdgeInput,
-  ): Promise<WorkflowVersionStepChangesDTO> {
-    return this.workflowVersionStepWorkspaceService.createWorkflowVersionEdge({
-      source,
-      target,
-      workflowVersionId,
-      workspaceId,
-    });
-  }
-
-  @Mutation(() => WorkflowVersionStepChangesDTO)
-  async deleteWorkflowVersionEdge(
-    @AuthWorkspace() { id: workspaceId }: Workspace,
-    @Args('input')
-    { source, target, workflowVersionId }: CreateWorkflowVersionEdgeInput,
-  ): Promise<WorkflowVersionStepChangesDTO> {
-    return this.workflowVersionStepWorkspaceService.deleteWorkflowVersionEdge({
-      source,
-      target,
-      workflowVersionId,
-      workspaceId,
-    });
   }
 }
