+++ packages/twenty-server/src/engine/core-modules/workflow/workflow-api.module.ts
@@ -3,7 +3,7 @@ import { Module } from '@nestjs/common';
 import { FeatureFlagModule } from 'src/engine/core-modules/feature-flag/feature-flag.module';
 import { WorkflowTriggerController } from 'src/engine/core-modules/workflow/controllers/workflow-trigger.controller';
 import { WorkflowBuilderResolver } from 'src/engine/core-modules/workflow/resolvers/workflow-builder.resolver';
-import { WorkflowStepResolver } from 'src/engine/core-modules/workflow/resolvers/workflow-step.resolver';
+import { WorkflowVersionStepResolver } from 'src/engine/core-modules/workflow/resolvers/workflow-version-step.resolver';
 import { WorkflowTriggerResolver } from 'src/engine/core-modules/workflow/resolvers/workflow-trigger.resolver';
 import { WorkflowVersionResolver } from 'src/engine/core-modules/workflow/resolvers/workflow-version.resolver';
 import { PermissionsModule } from 'src/engine/metadata-modules/permissions/permissions.module';
@@ -12,6 +12,7 @@ import { WorkflowBuilderModule } from 'src/modules/workflow/workflow-builder/wor
 import { WorkflowVersionModule } from 'src/modules/workflow/workflow-builder/workflow-version/workflow-version.module';
 import { WorkflowRunModule } from 'src/modules/workflow/workflow-runner/workflow-run/workflow-run.module';
 import { WorkflowTriggerModule } from 'src/modules/workflow/workflow-trigger/workflow-trigger.module';
+import { WorkflowVersionEdgeResolver } from 'src/engine/core-modules/workflow/resolvers/workflow-version-edge.resolver';
 
 @Module({
   imports: [
@@ -27,7 +28,8 @@ import { WorkflowTriggerModule } from 'src/modules/workflow/workflow-trigger/wor
   providers: [
     WorkflowTriggerResolver,
     WorkflowBuilderResolver,
-    WorkflowStepResolver,
+    WorkflowVersionStepResolver,
+    WorkflowVersionEdgeResolver,
     WorkflowVersionResolver,
   ],
 })
