+++ packages/twenty-server/src/modules/workflow/common/exceptions/workflow-version-edge.exception.ts
@@ -1,0 +1,8 @@
+import { CustomException } from 'src/utils/custom-exception';
+
+export class WorkflowVersionEdgeException extends CustomException<WorkflowVersionEdgeExceptionCode> {}
+
+export enum WorkflowVersionEdgeExceptionCode {
+  UNKNOWN = 'UNKNOWN',
+  NOT_FOUND = 'NOT_FOUND',
+}
