+++ packages/twenty-server/src/modules/workflow/workflow-builder/workflow-builder.module.ts
@@ -6,18 +6,21 @@ import { ObjectMetadataEntity } from 'src/engine/metadata-modules/object-metadat
 import { WorkflowSchemaModule } from 'src/modules/workflow/workflow-builder/workflow-schema/workflow-schema.module';
 import { WorkflowVersionStepModule } from 'src/modules/workflow/workflow-builder/workflow-version-step/workflow-version-step.module';
 import { WorkflowVersionModule } from 'src/modules/workflow/workflow-builder/workflow-version/workflow-version.module';
+import { WorkflowVersionEdgeModule } from 'src/modules/workflow/workflow-builder/workflow-version-edge/workflow-version-edge.module';
 
 @Module({
   imports: [
     WorkflowSchemaModule,
     WorkflowVersionModule,
     WorkflowVersionStepModule,
+    WorkflowVersionEdgeModule,
     NestjsQueryTypeOrmModule.forFeature([ObjectMetadataEntity], 'core'),
   ],
   exports: [
     WorkflowSchemaModule,
     WorkflowVersionModule,
     WorkflowVersionStepModule,
+    WorkflowVersionEdgeModule,
   ],
 })
 export class WorkflowBuilderModule {}
