+++ packages/twenty-server/src/modules/workflow/workflow-builder/workflow-version-step/workflow-version-step.workspace-service.ts
@@ -13,7 +13,6 @@ import { CreateWorkflowVersionStepInput } from 'src/engine/core-modules/workflow
 import { AgentService } from 'src/engine/metadata-modules/agent/agent.service';
 import { ObjectMetadataEntity } from 'src/engine/metadata-modules/object-metadata/object-metadata.entity';
 import { ServerlessFunctionService } from 'src/engine/metadata-modules/serverless-function/serverless-function.service';
-import { ScopedWorkspaceContextFactory } from 'src/engine/twenty-orm/factories/scoped-workspace-context.factory';
 import { TwentyORMGlobalManager } from 'src/engine/twenty-orm/twenty-orm-global.manager';
 import {
   WorkflowVersionStepException,
@@ -35,7 +34,7 @@ import { WorkflowRunWorkspaceService } from 'src/modules/workflow/workflow-runne
 import { WorkflowRunnerWorkspaceService } from 'src/modules/workflow/workflow-runner/workspace-services/workflow-runner.workspace-service';
 import { WorkflowStepPositionInput } from 'src/engine/core-modules/workflow/dtos/update-workflow-step-position-input.dto';
 import { WorkflowVersionStepChangesDTO } from 'src/engine/core-modules/workflow/dtos/workflow-version-step-changes.dto';
-import { computeWorkflowVersionStepChanges } from 'src/modules/workflow/workflow-builder/workflow-version-step/utils/compute-workflow-version-step-updates.util';
+import { computeWorkflowVersionStepChanges } from 'src/modules/workflow/workflow-builder/utils/compute-workflow-version-step-updates.util';
 
 const BASE_STEP_DEFINITION: BaseWorkflowActionSettings = {
   outputSchema: {},
@@ -61,7 +60,6 @@ export class WorkflowVersionStepWorkspaceService {
     private readonly workflowRunWorkspaceService: WorkflowRunWorkspaceService,
     private readonly workflowRunnerWorkspaceService: WorkflowRunnerWorkspaceService,
     private readonly workflowCommonWorkspaceService: WorkflowCommonWorkspaceService,
-    private readonly scopedWorkspaceContextFactory: ScopedWorkspaceContextFactory,
   ) {}
 
   async createWorkflowVersionStep({
@@ -365,244 +363,6 @@ export class WorkflowVersionStepWorkspaceService {
       workspaceId,
       workflowRunId,
       lastExecutedStepId: stepId,
-    });
-  }
-
-  async createWorkflowVersionEdge({
-    source,
-    target,
-    workflowVersionId,
-    workspaceId,
-  }: {
-    source: string;
-    target: string;
-    workflowVersionId: string;
-    workspaceId: string;
-  }): Promise<WorkflowVersionStepChangesDTO> {
-    const workflowVersionRepository =
-      await this.twentyORMGlobalManager.getRepositoryForWorkspace<WorkflowVersionWorkspaceEntity>(
-        workspaceId,
-        'workflowVersion',
-        { shouldBypassPermissionChecks: true },
-      );
-
-    const workflowVersion = await workflowVersionRepository.findOne({
-      where: {
-        id: workflowVersionId,
-      },
-    });
-
-    if (!isDefined(workflowVersion)) {
-      throw new WorkflowVersionStepException(
-        'WorkflowVersion not found',
-        WorkflowVersionStepExceptionCode.NOT_FOUND,
-      );
-    }
-
-    assertWorkflowVersionIsDraft(workflowVersion);
-
-    const steps = workflowVersion.steps || [];
-
-    const trigger = workflowVersion.trigger;
-
-    const isSourceTrigger = source === 'trigger';
-
-    const targetStep = steps.find((step) => step.id === target);
-
-    if (!isDefined(targetStep)) {
-      throw new WorkflowVersionStepException(
-        `Target step '${target}' not found in workflowVersion '${workflowVersionId}'`,
-        WorkflowVersionStepExceptionCode.NOT_FOUND,
-      );
-    }
-
-    if (isSourceTrigger) {
-      if (!isDefined(trigger)) {
-        throw new WorkflowVersionStepException(
-          `Trigger not found in workflowVersion '${workflowVersionId}'`,
-          WorkflowVersionStepExceptionCode.NOT_FOUND,
-        );
-      }
-
-      if (trigger.nextStepIds?.includes(target)) {
-        return computeWorkflowVersionStepChanges({
-          trigger,
-          steps,
-        });
-      }
-
-      const updatedTrigger = {
-        ...trigger,
-        nextStepIds: [...(trigger.nextStepIds ?? []), target],
-      };
-
-      await workflowVersionRepository.update(workflowVersion.id, {
-        trigger: updatedTrigger,
-      });
-
-      return computeWorkflowVersionStepChanges({
-        trigger: updatedTrigger,
-        steps,
-      });
-    }
-
-    const sourceStep = steps.find((step) => step.id === source);
-
-    if (!isDefined(sourceStep)) {
-      throw new WorkflowVersionStepException(
-        `Source step '${source}' not found in workflowVersion '${workflowVersionId}'`,
-        WorkflowVersionStepExceptionCode.NOT_FOUND,
-      );
-    }
-
-    if (sourceStep.nextStepIds?.includes(target)) {
-      return computeWorkflowVersionStepChanges({
-        trigger,
-        steps,
-      });
-    }
-
-    const updatedSourceStep = {
-      ...sourceStep,
-      nextStepIds: [...(sourceStep.nextStepIds ?? []), target],
-    };
-
-    const updatedSteps = steps.map((step) => {
-      if (step.id === source) {
-        return updatedSourceStep;
-      }
-
-      return step;
-    });
-
-    await workflowVersionRepository.update(workflowVersion.id, {
-      steps: updatedSteps,
-    });
-
-    return computeWorkflowVersionStepChanges({
-      trigger,
-      steps: updatedSteps,
-    });
-  }
-
-  async deleteWorkflowVersionEdge({
-    source,
-    target,
-    workflowVersionId,
-    workspaceId,
-  }: {
-    source: string;
-    target: string;
-    workflowVersionId: string;
-    workspaceId: string;
-  }): Promise<WorkflowVersionStepChangesDTO> {
-    const workflowVersionRepository =
-      await this.twentyORMGlobalManager.getRepositoryForWorkspace<WorkflowVersionWorkspaceEntity>(
-        workspaceId,
-        'workflowVersion',
-        { shouldBypassPermissionChecks: true },
-      );
-
-    const workflowVersion = await workflowVersionRepository.findOne({
-      where: {
-        id: workflowVersionId,
-      },
-    });
-
-    if (!isDefined(workflowVersion)) {
-      throw new WorkflowVersionStepException(
-        'WorkflowVersion not found',
-        WorkflowVersionStepExceptionCode.NOT_FOUND,
-      );
-    }
-
-    assertWorkflowVersionIsDraft(workflowVersion);
-
-    const steps = workflowVersion.steps || [];
-
-    const trigger = workflowVersion.trigger;
-
-    const isSourceTrigger = source === 'trigger';
-
-    const targetStep = steps.find((step) => step.id === target);
-
-    if (!isDefined(targetStep)) {
-      throw new WorkflowVersionStepException(
-        `Target step '${target}' not found in workflowVersion '${workflowVersionId}'`,
-        WorkflowVersionStepExceptionCode.NOT_FOUND,
-      );
-    }
-
-    if (isSourceTrigger) {
-      if (!isDefined(trigger)) {
-        throw new WorkflowVersionStepException(
-          `Trigger not found in workflowVersion '${workflowVersionId}'`,
-          WorkflowVersionStepExceptionCode.NOT_FOUND,
-        );
-      }
-
-      if (!trigger.nextStepIds?.includes(target)) {
-        return computeWorkflowVersionStepChanges({
-          trigger,
-          steps,
-        });
-      }
-
-      const updatedTrigger = {
-        ...trigger,
-        nextStepIds: trigger.nextStepIds?.filter(
-          (nextStepId) => nextStepId !== target,
-        ),
-      };
-
-      await workflowVersionRepository.update(workflowVersion.id, {
-        trigger: updatedTrigger,
-      });
-
-      return computeWorkflowVersionStepChanges({
-        trigger: updatedTrigger,
-        steps,
-      });
-    }
-
-    const sourceStep = steps.find((step) => step.id === source);
-
-    if (!isDefined(sourceStep)) {
-      throw new WorkflowVersionStepException(
-        `Source step '${source}' not found in workflowVersion '${workflowVersionId}'`,
-        WorkflowVersionStepExceptionCode.NOT_FOUND,
-      );
-    }
-
-    if (!sourceStep.nextStepIds?.includes(target)) {
-      return computeWorkflowVersionStepChanges({
-        trigger,
-        steps,
-      });
-    }
-
-    const updatedSourceStep = {
-      ...sourceStep,
-      nextStepIds: sourceStep.nextStepIds?.filter(
-        (nextStepId) => nextStepId !== target,
-      ),
-    };
-
-    const updatedSteps = steps.map((step) => {
-      if (step.id === source) {
-        return updatedSourceStep;
-      }
-
-      return step;
-    });
-
-    await workflowVersionRepository.update(workflowVersion.id, {
-      steps: updatedSteps,
-    });
-
-    return computeWorkflowVersionStepChanges({
-      trigger,
-      steps: updatedSteps,
     });
   }
